+++ app/schemas/impuestos.py
@@ -1,10 +1,11 @@
 from sqlmodel import SQLModel, Field
 from typing import Optional
+from decimal import Decimal
 
 class Impuesto(SQLModel):
     codigo_sri: str
     nombre: str
-    porcentaje: float
+    porcentaje: Decimal
     descripcion: str
     activo: bool = True
 
@@ -18,7 +19,7 @@ class ImpuestoRead(Impuesto):
 class ImpuestoUpdate(SQLModel):
     codigo_sri: Optional[str] = None
     nombre: Optional[str] = None
-    porcentaje: Optional[float] = None
+    porcentaje: Optional[Decimal] = None
     descripcion: Optional[str] = None
     activo: Optional[bool] = None
